requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning and data mining
scikit-learn>=1.3.0
mlxtend>=0.22.0
apyori>=1.1.2

# Network analysis
networkx>=3.0
igraph>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Data formats
jsonschema>=4.17.0
pyyaml>=6.0

# Performance
numba>=0.57.0
joblib>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

